move getStackTrace back to DisplayUtils
