+++ src/main/java/org/sagebionetworks/web/client/ClientLoggerImpl.java
@@ -2,8 +2,6 @@ package org.sagebionetworks.web.client;
 
 import java.io.StringWriter;
 
-import org.sagebionetworks.web.client.widget.entity.controller.SynapseAlertImpl;
-
 import com.google.gwt.user.client.rpc.AsyncCallback;
 import com.google.inject.Inject;
 
@@ -88,7 +86,7 @@ public class ClientLoggerImpl implements ClientLogger{
 	public void errorToRepositoryServices(String message, Throwable e){
 		String exceptionName = e != null ? e.getClass().getName() : "";
 		String exceptionMessage = e != null && e.getMessage() != null ? e.getMessage() : "";
-		String stackTrace = SynapseAlertImpl.getStackTrace(e); 
+		String stackTrace = DisplayUtils.getStackTrace(e); 
 		this.synapseClient.logErrorToRepositoryServices(message + "-" + exceptionMessage + "-" + stackTrace, exceptionName, new AsyncCallback<Void>() {
 			@Override
 			public void onSuccess(Void result) {
