+++ src/main/java/org/sagebionetworks/web/client/widget/entity/JiraURLHelperImpl.java
@@ -8,7 +8,6 @@ import org.sagebionetworks.web.client.DisplayUtils;
 import org.sagebionetworks.web.client.GWTWrapper;
 import org.sagebionetworks.web.client.JiraClientAsync;
 import org.sagebionetworks.web.client.security.AuthenticationController;
-import org.sagebionetworks.web.client.widget.entity.controller.SynapseAlertImpl;
 
 import com.google.gwt.user.client.Window.Navigator;
 import com.google.gwt.user.client.rpc.AsyncCallback;
@@ -209,7 +208,7 @@ public class JiraURLHelperImpl implements JiraURLHelper {
 			AsyncCallback<Void> callback) {
 		Map<String, String> fieldValues = new HashMap<String, String>();
 		
-		String stackTrace = SynapseAlertImpl.getStackTrace(t);
+		String stackTrace = DisplayUtils.getStackTrace(t);
 		
 		StringBuilder description = new StringBuilder();
