+++ src/main/java/org/sagebionetworks/web/client/widget/provenance/ProvenanceWidget.java
@@ -27,7 +27,6 @@ import org.sagebionetworks.web.client.security.AuthenticationController;
 import org.sagebionetworks.web.client.transform.JsoProvider;
 import org.sagebionetworks.web.client.utils.Callback;
 import org.sagebionetworks.web.client.widget.WidgetRendererPresenter;
-import org.sagebionetworks.web.client.widget.entity.controller.SynapseAlertImpl;
 import org.sagebionetworks.web.client.widget.provenance.nchart.LayoutResult;
 import org.sagebionetworks.web.client.widget.provenance.nchart.NChartCharacters;
 import org.sagebionetworks.web.client.widget.provenance.nchart.NChartLayersArray;
@@ -509,7 +508,7 @@ public class ProvenanceWidget implements ProvenanceWidgetView.Presenter, WidgetR
 			public void onFailure(Throwable caught) {
 				if (caught instanceof NotFoundException) {
 					//SWC-1843: do not redirect home.  log full exception to the console
-					synapseJSNIUtils.consoleError(caught.getMessage() + "\n" + SynapseAlertImpl.getStackTrace(caught));
+					synapseJSNIUtils.consoleError(caught.getMessage() + "\n" + DisplayUtils.getStackTrace(caught));
 				} else {
 					DisplayUtils.handleServiceException(caught, globalApplicationState, authenticationController.isLoggedIn(), view);	
 				}
